backend_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.8.1
spacy==3.7.5
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.32.3
python-multipart==0.0.12
python-dotenv==1.0.1
